refactor: extract UI toggle logic into shared utility with error handling
Address review feedback from Copilot and CodeRabbit:
- Extract duplicated hideUserInterface/toggleNavigationPanel/enableRestMode
  logic into a shared ui-toggle.ts utility (DRY principle)
- Add try/catch error handling with console.error prefix per CLAUDE.md
- Add bounds check for children[1] access on navigator-toolbox
